+++ contracts/src/token/erc1155/extensions/metadata_uri.rs
@@ -8,7 +8,7 @@ use alloc::{string::String, vec, vec::Vec};
 use alloy_primitives::{FixedBytes, U256};
 use openzeppelin_stylus_proc::interface_id;
 pub use sol::*;
-use stylus_sdk::{prelude::*, storage::StorageString, stylus_proc::public};
+use stylus_sdk::{prelude::*, storage::StorageString};
 
 use crate::utils::introspection::erc165::{Erc165, IErc165};